refactor: migrate client to Feature-Sliced Architecture
- Move components and hooks into features/ directory structure:
  - features/auth/ — login-form, signup-form, header, use-auth
  - features/entries/ — entry-card, entry-editor, entry-list, use-entries, use-entry
  - features/questions/ — question-card, use-questions
- Extract API logic from page components into dedicated hooks:
  - useEntries: cursor-based pagination for entry list
  - useEntry/useSaveEntry/useDeleteEntry: single entry CRUD
  - useQuestions: question list fetch and creation
- Remove `as` type casts from entry hooks (use type narrowing)
- Update all page imports to use @/features/ paths
- Delete old components/ and hooks/ directories
