+++ apps/client/src/app/(auth)/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from '@/components/auth/login-form';
+import { LoginForm } from '@/features/auth/components/login-form';
 
 export default function LoginPage() {
   return <LoginForm />;
