+++ apps/client/src/app/(auth)/signup/page.tsx
@@ -1,4 +1,4 @@
-import { SignupForm } from '@/components/auth/signup-form';
+import { SignupForm } from '@/features/auth/components/signup-form';
 
 export default function SignupPage() {
   return <SignupForm />;
