+++ apps/client/src/app/(protected)/entries/[id]/page.tsx
@@ -1,46 +1,24 @@
 'use client';
 
 import { useParams, useRouter } from 'next/navigation';
-import { useEffect, useState } from 'react';
-import { EntryEditor } from '@/components/entries/entry-editor';
-import { useAuth } from '@/hooks/use-auth';
-
-interface EntryDetail {
-  entry: {
-    id: string;
-    content: string;
-    createdAt: string;
-    updatedAt: string;
-  };
-}
+import { useAuth } from '@/features/auth/hooks/use-auth';
+import { EntryEditor } from '@/features/entries/components/entry-editor';
+import { useDeleteEntry, useEntry } from '@/features/entries/hooks/use-entry';
 
 export default function EntryDetailPage() {
   const { id } = useParams<{ id: string }>();
-  const { api, loading: authLoading } = useAuth();
-  const [entry, setEntry] = useState<EntryDetail['entry'] | null>(null);
-  const [loading, setLoading] = useState(true);
-  const [deleting, setDeleting] = useState(false);
+  const { api, auth, loading: authLoading } = useAuth();
+  const { entry, loading } = useEntry(id, api, authLoading);
+  const { remove, deleting } = useDeleteEntry(api);
   const router = useRouter();
 
-  useEffect(() => {
-    if (authLoading || !api) return;
-
-    api.api.v1.entries[':id'].$get({ param: { id } }).then(async (res) => {
-      if (res.ok) {
-        const data = await res.json();
-        const e = 'entry' in data ? (data as EntryDetail).entry : data;
-        setEntry(e as EntryDetail['entry']);
-      }
-      setLoading(false);
-    });
-  }, [api, authLoading, id]);
-
   async function handleDelete() {
-    if (!api || !confirm('このエントリを削除しますか？')) return;
-    setDeleting(true);
-    await api.api.v1.entries[':id'].$delete({ param: { id } });
-    router.push('/entries');
-    router.refresh();
+    if (!confirm('このエントリを削除しますか？')) return;
+    const ok = await remove(id);
+    if (ok) {
+      router.push('/entries');
+      router.refresh();
+    }
   }
 
   if (loading || authLoading) {
@@ -64,7 +42,7 @@ export default function EntryDetailPage() {
           {deleting ? '削除中...' : '削除'}
         </button>
       </div>
-      <EntryEditor entryId={entry.id} initialContent={entry.content} />
+      <EntryEditor entryId={entry.id} initialContent={entry.content} api={api} auth={auth} />
     </div>
   );
 }
