+++ apps/client/src/app/(protected)/entries/new/page.tsx
@@ -1,10 +1,15 @@
-import { EntryEditor } from '@/components/entries/entry-editor';
+'use client';
+
+import { useAuth } from '@/features/auth/hooks/use-auth';
+import { EntryEditor } from '@/features/entries/components/entry-editor';
 
 export default function NewEntryPage() {
+  const { api, auth } = useAuth();
+
   return (
     <div className="flex flex-col gap-4">
       <h1 className="text-xl font-bold">新しいエントリ</h1>
-      <EntryEditor />
+      <EntryEditor api={api} auth={auth} />
     </div>
   );
 }
