+++ apps/client/src/features/auth/hooks/use-auth.ts
