+++ apps/client/src/features/entries/components/entry-card.tsx
